Move OutputTab and remove auto-gen comments
