+++ src/main/java/org/bric/gui/BricUI.java
@@ -6,8 +6,12 @@ package org.bric.gui;
 
 import org.bric.core.model.ImportedImage;
 import org.bric.gui.inputOutput.ProgressBarFrame;
+import org.bric.gui.output.OutputTab;
 import org.bric.gui.preferences.PreferencesFrame;
-import org.bric.gui.tabs.*;
+import org.bric.gui.tabs.ImageEditTab;
+import org.bric.gui.tabs.ResizeJPanel;
+import org.bric.gui.tabs.RotateJPanel;
+import org.bric.gui.tabs.WatermarkJPanel;
 import org.bric.imageEditParameters.*;
 import org.bric.processor.ImageProcessHandler;
 import org.bric.utils.ArrayListTransferHandler;
@@ -68,7 +72,7 @@ public class BricUI extends JFrame {
         inputList.setModel(model);
         inputList.setTransferHandler(arrayListHandler);
         inputList.setDragEnabled(true);
-        editPane.add(bundle.getString("BricUI.outputTab.name"), new OutputJPanel());
+        editPane.add(bundle.getString("BricUI.outputTab.name"), new OutputTab());
         editPane.add(bundle.getString("BricUI.resizeTab.name"), new ResizeJPanel());
         editPane.add(bundle.getString("BricUI.rotateTab.name"), new RotateJPanel());
         editPane.add(bundle.getString("BricUI.watermarkTab.name"), new WatermarkJPanel());
@@ -874,11 +878,11 @@ public class BricUI extends JFrame {
             }
             out = new FileOutputStream(new File(file));
             
-            OutputJPanel outputJPanel = (OutputJPanel) editPane.getComponentAt(0);
-            properties.setProperty("fileTypeCombo", Integer.toString(outputJPanel.getFileTypeComboIndex()));
-            properties.setProperty("outputPathText", outputJPanel.getOutputPathText());
-            properties.setProperty("qualityValue", Integer.toString(outputJPanel.getQualitySliderValue()));
-            properties.setProperty("startIndexValue", Integer.toString(outputJPanel.getStartIndexSpinnerValue()));
+            OutputTab outputTab = (OutputTab) editPane.getComponentAt(0);
+            properties.setProperty("fileTypeCombo", Integer.toString(outputTab.getFileTypeComboIndex()));
+            properties.setProperty("outputPathText", outputTab.getOutputPathText());
+            properties.setProperty("qualityValue", Integer.toString(outputTab.getQualitySliderValue()));
+            properties.setProperty("startIndexValue", Integer.toString(outputTab.getStartIndexSpinnerValue()));
             
             ResizeJPanel resizeJPanel = (ResizeJPanel) editPane.getComponentAt(1);
             properties.setProperty("resizeAntialising", resizeJPanel.getAntialisingCheckBox() == true ? "1" : "0");
@@ -941,11 +945,11 @@ public class BricUI extends JFrame {
             fileInput = new FileInputStream(file);
             properties.load(fileInput);
             
-            OutputJPanel outputJPanel = (OutputJPanel) editPane.getComponentAt(0);
-            outputJPanel.setFileTypeComboIndex(properties.getProperty("fileTypeCombo"));
-            outputJPanel.setOutputPathText(properties.getProperty("outputPathText"));
-            outputJPanel.setQualitySliderValue(properties.getProperty("qualityValue"));
-            outputJPanel.setStartIndexSpinnerValue(Integer.parseInt(properties.getProperty("startIndexValue")));
+            OutputTab outputTab = (OutputTab) editPane.getComponentAt(0);
+            outputTab.setFileTypeComboIndex(properties.getProperty("fileTypeCombo"));
+            outputTab.setOutputPathText(properties.getProperty("outputPathText"));
+            outputTab.setQualitySliderValue(properties.getProperty("qualityValue"));
+            outputTab.setStartIndexSpinnerValue(Integer.parseInt(properties.getProperty("startIndexValue")));
             
             ResizeJPanel resizeJPanel = (ResizeJPanel) editPane.getComponentAt(1);
             resizeJPanel.setAntialisingCheckBox(properties.getProperty("resizeAntialising").equals("1") ? true : false);
