+++ src/main/java/org/bric/gui/output/OutputTab.java
@@ -1,28 +1,34 @@
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
-package org.bric.gui.tabs;
+package org.bric.gui.output;
 
 import org.bric.gui.BricUI;
+import org.bric.gui.tabs.ImageEditTab;
 import org.bric.imageEditParameters.OutputParameters;
 import org.bric.utils.Utils;
 
 import javax.swing.*;
 import java.util.ResourceBundle;
 
-/**
- *
- * @author vasilis
- */
-public class OutputJPanel extends javax.swing.JPanel implements ImageEditTab{
-    
+public class OutputTab extends javax.swing.JPanel implements ImageEditTab {
+
+    private javax.swing.JButton browseButton;
+    private javax.swing.JLabel dpiLabel;
+    private javax.swing.JTextPane fileNameMasksPane;
+    private javax.swing.JScrollPane fileNameMasksScrollPane;
+    private javax.swing.JComboBox fileTypeCombo;
+    private javax.swing.JLabel formatLabel;
+    private javax.swing.JLabel jLabel7;
+    private javax.swing.JLabel jLabel8;
+    private javax.swing.JSpinner jSpinner2;
+    private javax.swing.JSpinner jSpinner3;
+    private javax.swing.JLabel numberingLabel;
+    private javax.swing.JTextField outputPathText;
+    private javax.swing.JLabel qualityLabel;
+    private javax.swing.JSlider qualitySlider;
+    private javax.swing.JSpinner startIndexSpinner;
+
     ResourceBundle bundle;
-    
-    /**
-     * Creates new form OutputJPanel
-     */
-    public OutputJPanel() {
+
+    public OutputTab() {
         
         bundle = ResourceBundle.getBundle("lang/gui/tabs/OutputJPanel");
         
@@ -43,13 +49,6 @@ public class OutputJPanel extends javax.swing.JPanel implements ImageEditTab{
         jSpinner3.setVisible(false);
     }
 
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
     private void initComponents() {
 
         outputPathText = new javax.swing.JTextField();
@@ -67,8 +66,6 @@ public class OutputJPanel extends javax.swing.JPanel implements ImageEditTab{
         jLabel7 = new javax.swing.JLabel();
         jLabel8 = new javax.swing.JLabel();
         jSpinner3 = new javax.swing.JSpinner();
-
-        //setBackground(new java.awt.Color(43, 43, 43));
 
         outputPathText.setText(System.getProperty("user.home"));
 
@@ -184,7 +181,7 @@ public class OutputJPanel extends javax.swing.JPanel implements ImageEditTab{
                 .addGap(18, 18, 18)
                 .addComponent(fileNameMasksScrollPane, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
         );
-    }// </editor-fold>//GEN-END:initComponents
+    }
 
     private void browseButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_browseButtonActionPerformed
         JFileChooser chooser = new JFileChooser(BricUI.lastOpenedDirectory);
@@ -194,7 +191,7 @@ public class OutputJPanel extends javax.swing.JPanel implements ImageEditTab{
         }
         outputPathText.setText(chooser.getSelectedFile().getPath() + Utils.FS);
         BricUI.lastOpenedDirectory = chooser.getSelectedFile().getPath();
-    }//GEN-LAST:event_browseButtonActionPerformed
+    }
 
     private void fileTypeComboItemStateChanged(java.awt.event.ItemEvent evt) {//GEN-FIRST:event_fileTypeComboItemStateChanged
         boolean qualityEnable = fileTypeCombo.getSelectedIndex() <= 2 ? true : false;
@@ -205,29 +202,11 @@ public class OutputJPanel extends javax.swing.JPanel implements ImageEditTab{
         jLabel8.setEnabled(qualityEnable);
         jSpinner2.setEnabled(qualityEnable);
         jSpinner3.setEnabled(qualityEnable);
-    }//GEN-LAST:event_fileTypeComboItemStateChanged
+    }
 
     private void qualitySliderStateChanged(javax.swing.event.ChangeEvent evt) {//GEN-FIRST:event_qualitySliderStateChanged
         qualityLabel.setText(String.format(bundle.getString("OutputJPanel.qualityLabel.text") + "%s%%", qualitySlider.getValue()));
-    }//GEN-LAST:event_qualitySliderStateChanged
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JButton browseButton;
-    private javax.swing.JLabel dpiLabel;
-    private javax.swing.JTextPane fileNameMasksPane;
-    private javax.swing.JScrollPane fileNameMasksScrollPane;
-    private javax.swing.JComboBox fileTypeCombo;
-    private javax.swing.JLabel formatLabel;
-    private javax.swing.JLabel jLabel7;
-    private javax.swing.JLabel jLabel8;
-    private javax.swing.JSpinner jSpinner2;
-    private javax.swing.JSpinner jSpinner3;
-    private javax.swing.JLabel numberingLabel;
-    private javax.swing.JTextField outputPathText;
-    private javax.swing.JLabel qualityLabel;
-    private javax.swing.JSlider qualitySlider;
-    private javax.swing.JSpinner startIndexSpinner;
-    // End of variables declaration//GEN-END:variables
+    }
 
     public int getFileTypeComboIndex() {
         return fileTypeCombo.getSelectedIndex();
@@ -268,8 +247,6 @@ public class OutputJPanel extends javax.swing.JPanel implements ImageEditTab{
         outputParameters.setQuality((float)qualitySlider.getValue()/100);
         outputParameters.setOutputFormat(fileTypeCombo.getSelectedItem().toString());
         outputParameters.setNumberingStartIndex((Integer)startIndexSpinner.getValue());
-//        outputParameters.setDpiHorizontal((Integer) jSpinner2.getValue());
-//        outputParameters.setDpiVertical((Integer) jSpinner3.getValue());
         return outputParameters;
     }
 }
